# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Landly API
  description: API для платформы генерации AI-лендингов
  version: 1.0.0
  contact:
    email: support@landly.ru
servers:
  - url: http://localhost:8080/v1
    description: Development server (API)
  - url: https://api.landly.ru/v1
    description: Production server (API)
  - url: http://localhost:8080
    description: Development server (Sites)
  - url: https://landlify.ru
    description: Production server (Sites)
...
paths:
  /auth/signup:
    post:
      tags: [auth]
      summary: Регистрация нового пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
      responses:
        '200':
          description: Успешная регистрация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /auth/login:
    post:
      tags: [auth]
      summary: Вход в систему
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Успешный вход
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
...
  /analytics/{id}/event:
    post:
      tags: [analytics]
      summary: Отправить событие аналитики
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [event_type, path]
              properties:
                event_type:
                  type: string
                  enum: [pageview, cta_click, pay_click, form_submit]
                path:
                  type: string
                referrer:
                  type: string
      responses:
        '204':
          description: Событие зарегистрировано
  /analytics/{id}/stats:
    get:
      tags: [analytics]
      summary: Получить статистику проекта
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Статистика
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsStats'

  # Published Sites (canonical URLs without /v1)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_at:
          type: string
          format: date-time
...
    AnalyticsStats:
      type: object
      properties:
        project_id:
          type: string
          format: uuid
        total_pageviews:
          type: integer
        total_cta_clicks:
          type: integer
        total_pay_clicks:
          type: integer
        unique_visitors:
          type: integer